Fix a longstanding crash bug with kqueue in 1.4
The problem occurs because queueing an add and a del in the same
kevent changes list for the same [event,filter] combination can result
in the event being reported if it was pending.  This is a big problem
for 1.4.x, especially if the del was followed up by freeing the event
in question!

This patch should fix that by only keeping the most recently added
event for any event,filter combination.  This will result in some
needless deletes, but we already have the code in place to ignore the
errors from those.

In 2.0, the evmap code already removes the crash condition, and the
event_changelist code handles coalescing redundant changes.

Thanks to Charles Kerr and various Transmission users for diagnosing
this.