Make the default evmap implementation into a flat array
In 2.0.1..2.0.4, the IO evmap on non-windows platforms was implemented as an
array of pointers to evmap_io, with a variable-sized fdinfo appended to each
evmap_io.  Thus, looking up an evmap_io for 'fd' required a pointer
dereference to find the pointer at map->entries[fd], a test to make sure it
was non-NULL, and another pointer dereference to use it.

This patch changes evmap->entries to a flat array.  Because we don't know the
size of the array's entries at compile time, we need to do array lookup
ourselves, but that isn't so hard.

Note that when we realloc the array, we need to fix up the event_lists in it
so that the correct prev pointer points to &tqh_first.