Avoid needless calls to ioctl(FIONREAD).
If we're asking for an amount of data that would fit cleanly in the
available space we're reading into, then we will never need to
reallocate the space we have, and so there's no need to check how much
data is waiting.

Based on code by Peter Szabo for Libevent 1.4.x.