Work around a kernel bug in kqueue when given a half-closed pipe()
The kevent call is supposed to report errors in its event list when
possible, not by returning -1 and setting errno.  But current versions
of OpenBSD, OSX, and FreeBSD all seem to do the wrong thing here when
given a pipe the other end of which has been closed: instead of
putting the error in the output, they have kevent() fail.

This patch detects that particular failure condition and responds by
acting on the changes one-by-one.

Diagnosed by Nicholas Marriott and Dale Rahn.