Avoid excessive timerfd_settime calls when the time hasn't changed
This is a start, but it's bogus: it needs to take into account that
we might get different times between epoll_dispatch calls not
because an event was added/removed, but because we were called with
differevent EVLOOP_ flags.

Can you tell that I plan to do some !squash commits on this branch?