Previous attempt to deal with queue.h failed.  Use a bigger hammer.
On *BSD systems, other headers may pull in <sys/queue.h>, and if this
happened in any file that included "queue-internal.h", we got floods
of multiply-defined macro errors.  The bigger hammer is to rename all
the macros that our copy of queue.h defines.  This means a lot of
mechanical changes throughout the code, but on the upside, there's no
longer any reason not to put the copy in include/event2/ and let the
various *_struct.h headers that want LIST_ENTRY or TAILQ_ENTRY use it;
this gets rid of some particularly nasty macro gunk.  (It is labeled
as internal-use-only and can be pulled back to an -internal.h when we
unexport the *_struct.h headers.)